Accept it when the configcook script path starts with the base directory.
Base directory is the directory where the configfile is,
or in the case of a remote file: the current directory.

Using 'python -m configcook' still does not work though,
or at least cannot find its own extension:

ERROR: We have package configcook but could not find a configcook.extension entrypoint with name configcook:extension_example.