Added note for pytest --pdb and pdb.post_mortem.
We could detect that we are in a pytest with --pdb,
but at this point we cannot know whether the sys.exit(1) is expected with pytest.raises(SystemExit) or not.
If it is expected, then we definitely would not want to start a pdb.
So simply add a note with the right command to use.